+++ CHANGELOG.md
@@ -1,5 +1,39 @@
-## [0.0.0] (2023-05-28)
+## [0.1.0-dev.3](https://github.com/javier-games/property-refs/compare/0.1.0-dev.2...0.1.0-dev.3) (2023-06-15)
 
-### ✨ Added
 
-* Initial commit.
+### :hammer: Changed
+
+* Property references view. ([6245b94](https://github.com/javier-games/property-refs/commit/6245b94811042f7d87c301d2876e44869e8c526a))
+* Samples to Samples~ ([59f05d0](https://github.com/javier-games/property-refs/commit/59f05d053615088c37de5051a996746b44d3dcdb))
+
+## [0.1.0-dev.2](https://github.com/javier-games/property-refs/compare/0.1.0-dev.1...0.1.0-dev.2) (2023-06-09)
+
+
+### :sparkles: Added
+
+* Auto code generation directive. ([db35c32](https://github.com/javier-games/property-refs/commit/db35c321e59b5b6699ab87111301933f6c3b187f))
+* Registry provider interface. ([5d3073c](https://github.com/javier-games/property-refs/commit/5d3073ca22148a9e3ee434a98f528568c4bacb06))
+* Roslyn source generator. ([c5e1c52](https://github.com/javier-games/property-refs/commit/c5e1c525cdc17b380e0fd1df0155c356acf9fa43))
+* Source generation sample. ([8ff4461](https://github.com/javier-games/property-refs/commit/8ff44617a326d55a033f712c926909d16620a75d))
+* Source generator to CI workflow. ([7d69001](https://github.com/javier-games/property-refs/commit/7d6900150cbb8561d7d43ad225f1cd5f0dd1418b))
+
+
+### :wrench: Fixed
+
+* IDE Missing reference error. ([2ecd3e6](https://github.com/javier-games/property-refs/commit/2ecd3e60f1ca58cbfccfaa4814d9b9207bd9f46a))
+* Property Ref Drawer and Registry file edition. ([36a9dc9](https://github.com/javier-games/property-refs/commit/36a9dc98f945503672251f666b4421d7b224c21a))
+
+## [0.1.0-dev.1](https://github.com/javier-games/property-refs/compare/0.0.0...0.1.0-dev.1) (2023-06-09)
+
+
+### :sparkles: Added
+
+* Property drawer. ([ac88ece](https://github.com/javier-games/property-refs/commit/ac88ece9906fe0e8522ebf6cfe9b52e44d7a9057))
+* Property refs and registry. ([815b378](https://github.com/javier-games/property-refs/commit/815b3783ddeb5872dc6cf5fefc61541403651390))
+* Refs component for prefabs sample. ([8cf3be0](https://github.com/javier-games/property-refs/commit/8cf3be003057f0017ff48c99d320b88f731ebffa))
+
+## [0.0.0](https://github.com/javier-games/property-refs/commit/f8dbb939a8372587f2d8c1d76127afebef169e1e) (2023-05-28)
+
+### :sparkles: Added 
+
+* Initial commit. ([f8dbb93](https://github.com/javier-games/property-refs/commit/f8dbb939a8372587f2d8c1d76127afebef169e1e))